Eliminate extra loop through modules and pass Zend\Module\Manager to module init()
Instead of looping through the modules first to run their init()'s, then again
to merge the configs, it now merges the configs during the initial loop through
the modules if the cache is off. Otherwise the cached config is loaded up front
and no loading or merging of module configs takes place.

Also, the per Matthew's suggestion, the Module Manager now passes itself to the
module's init() method to allow the module classes better access as to the state
of module loading, should they need it.