Create test to check if correct SM is used
The view helper can retrieve the container as a service from the service
manager. However, the view helper itself has a view helper service manager
and not the common Zend\ServiceManager\ServiceManager manager. This test
checks if the Zend\ServiceManager\ServiceManager is used to retrieve the
container when it is set.